replace the bit with a counter
The problem with the bit flip is that it's extremely unfair between
threads, as the one which owns a cache line may well flip it a lot
and allow only its sibling to benefit from two flips to perform its
change. With a counter it's easier to check for missed attempts.