use a shared runner number (ticket-like)
now all threads wait for their number to be called by looping on reads
on a separate variable that is isolated from the areas of operations.

The performance is slightly better in SMT (+4.5%) but slower with plain
cores (-5%).

Maybe a new test should be made with shared cache lines for same L1 caches.