store thread IDs in L1 cache and share them across threads of a same core
Now two threads from the same core watch the same runner ID. This increases
the perf by 5% thanks to faster takeover half of the time.

On the Xeon even when not sharing any CPU, this results in 88% performance
improvement (99 ns vs 187 ns). However on the Xeon with SMT it loses 3%.