Optionally remove the relax in the CAS
Now the performance increases by 28% on core i7+SMT with gmask 0. It
is possible that this happens because the two threads of a core perform
few operations in the loop and are not too much disturbed by their
neighbor, so the relax is counter-productive. Maybe with more operations
it would change but if the goal is to quickly pass a lock or an atomic
op to another thread then we don't need the relax.

On Xeon+SMT it's still 10% better than before (gmask 0 or 1). On
xeon without SMT it's 5-10% faster than before.

On a Ryzen 2700X, relax slightly improves the performance (+10% from
19.4 to 21.5M ops/s)

On Cortex A72, cpu relax has no effect.

On Graviton2, it slightly helps, but helps a bit more when combined
with grouping at level 1 (-r 1 -g 1 gives 100ns vs 110 for 64 threads,
and 92 vs 108 for 16 threads).

It seems that g1r1 also slightly improves the situation on pure cores
on x86.