Implement pure xchg with relax=2 and xchg+pause with relax=3
This will only be usable when tgmask=0. The xchg is cheaper than a CAS and
the performance further improves on core i7 (+2%). Same on the Xeon. xchg+pause
is bad on core i7 and xeon (the pause is always a problem there).

On the ryzen it gives the same performance as the cpu_relax.

On armv8-a it's 7.8% faster. And xchg+pause is 8.8% faster.

On v8.2-a with 64 cores it's either the same or slightly slower (~0-3% slower).
But it's 20% faster with 8 cores, 5% faster with 16 cores, and a bit slower
from 32 cores and above. However xchg+relax is always as good as CAS or better.