x86: use a shorter pause
Since we don't want to lose 30ns in a real pause instruction, let's just
perform a few register swaps. This is enough to leave the ALU available
for the other thread running on the same core without hurting the latency.

On both core i7 and xeon, it gives another 2.5% performance boost compared
to xchg with no pause. And xchg doesn't touch flags.