reimplement the load+store mode
Finally it works better on armv8, where it looks like the caches disturb
each other a lot. In mode 5 (load+store+isb), the a53 communication on
RK3399 drops from 50ns in mode 1 to 34 ns in mode 5. It's even 31 ns on
S5P6818. For the A72, it drops from 105ns to 94ns. The A53-A72 on RK3399
drops from 655ns to 477ns, and the big-to-little on S5P6818 drops from
372 to 279ns. So it looks clear that even the failed CAS on such
platforms have an important cost.