add a preliminary version of atomic_wrap()
This one is made of a bitmask and a pointer to the contention point.
However there seems to be at least one case where it doesn't stop, but
this can already happen if a lone thread fails last before others quit,
e.g. because it got the old value wrong.

Also the cost remains high. Another approach would be needed using
only a bitmask. The problem is that while a thread wakes another one,
a third one can pass before and succeed, forcing the second one to fail
and loop again.