add a supposedly better (but still broken) model
Now we have a wait phase before the operation and a wrap at the end. But
it's still unreliable, and is 5-8 times slower. There are too many cache
lines in use, and being forced to try the operation again is a disaster,
especially when it's almost certain to fail.

We need a better way to know that we can try the operation again or not.
In theory only the second thread ought to try the CAS again to check for
the original one's presence, but here we try to do this but there seems
to be corner cases. Also we have several TOCTOU issues that are always
related to the difference in time between the pointer and the bit.

Maybe another approach would be to let the second thread always loop but
subscribe. This would only cause some two-thread contention. But on NUMA
it would be catastrophic.