cas-lat: measure the latency needed to win a CAS
This measures the number of times a successful and a failed CAS happen
after a given CPU was just seen on it, and counts the number of times
more than 256 consecutive failures happen ("l0"), 4096 ("l1"), 65536
("l2"), 1048576 ("l4").

The result on Xeon are very fair, but on ryzen they're very unfair and
it's clearly visible that each group of 4 cores has an unfair access to
the local L3. Using exactly 5 xchg as a cpu_relax significantly mitigates
this and improves the overall loop rate.