cas-lat: there's still a race when leaving
A new waiter may not know the old runner just left. This is visible with
clang, or when adding "if (ticket < 0)" in the first loop with gcc. Since
we can't afford to perform a second CAS on output, better keep a little
bit of contention maybe, so that the first contender may always continue
to try ?