cas-lat: implement exp backoff
We limit the wait loop to 16M. This reduces the average cycle time to ~14ns,
or around 1/8 of what it was, but with L3 around 50. Smaller loop limits
improve the situation a little bit regarding max delays and limits performance
a bit. Using some variants of cpu_relax also helps with peak latencies but
decreases overall performance.