cas-lat: perform a load before the CAS to read the first value
otherwise we always fail one round. This saves ~30ns on core i7 and
nothing on AMD.

A better approach would be to accept to fail exactly one CAS to read
the value, to not re-read after a quick test, but to always read after
a long loop.

BROKEN: previous patches were broken, queue is never set to non-zero
thus we never wait. All the savings come from the delay incurred in
the jumps.