cas-bench: only update avg_wait once
Let's use a CAS instead of a store to update avg_wait. That way only
the first thread failing the race will update it, and it will not be
switched back from the small value to the large one after a successful
unlock. On the i7 this reduces the avg latency from 110 to 85ns. On the
Ryzen, it goes from 230 to 185 ns and the max almost never exceeds 44
attempts.