cas-bench: reduce the wait time less drastically on success
Instead of dividing it by 16, divide it by 4. The /16 used to give better
numbers because we were sometimes replacing zero with it. Now that we do
not replace zero, /4 is a better match and gives better numbers than /2 or
/8, especially on the Ryzen where the avg time is now 119 ns. It could
possibly have a relation with the number of cores per cache.