cas-bench: re-arrange the loop to have a preamble and epilogue
This construct will help have a preliminary check and an epilogue. However
it causes the avg_curr to always be updated, but despite this the performance
is better, both on i7 and Ryzen (101ns). Each attempt at not writing the first
value results in lower performance. It's possible that this is caused by branch
misprediction.