cas-bench: write small values upon each new attempt
pre-incrementing the counter helps reduce the average response time. By
writing 1 when entering we're down to 94ns. We can even go down to 66 by
writing 0 here and upon success. This proves that threads wait for too
long in the loop and need to compete a little bit more on the CAS. This
should be refined so that low values are not respected but large ones are.
This will make sure that long-waiting threads are served first but that
on limited congestion all threads can reasonably compete on the CAS.