cas-bench: place the waiting block at the beginning
We're relying on the loop count not being zero or the avg_curr not being
zero to enter the block, and we force bit0 of the loop count to 1 after
it to make sure we always get there (doing a ++ is less efficient). This
also allows the compiler to figure how to emit the code since it knows
the test on loopcnt will never fail after the first iteration.

With this we're down to 67 ns per CAS on 16 threads, with max attempts
almost never reaching 40.

An attempt to refresh avg_curr is made every time the loop count is twice
more than twice its value, and the value is re-read at least every 16
loops. This guarantees that waiters are quickly notified.

The loops time remains at 201 ns on Armada8040 (2 dual-core clusters).