cas-bench: don't wait unless really needed
On armv8.2 the current serialization triples the wait time. By applying
careful checks it's possible to avoid serializing if the contention is
low. Repeated tests on different platforms have shown that serializing
only with avg_curr>2 or failcnt>4 is optimal. This slightly increases
x86 time from 63 to 65 ns but most often prevents arm from having to
serialize if not needed, remaining at 35ns instead of 96 previously. On
armada8040 with two clusters, the timing is slightly lowered to 196ns
as well.