cas-bench: try not to use failcnt in the first block
We can rearrange the loop to only rely on loopcnt with
arbitrary values. Certain offsets allow to reach 59-60 ns, same for the
value that is written back, but these become delicate to tune. It's possible
to go further but then i7 goes out of tune, this will be done in a different
mode.