cas-bench: commit experimental relax mode 2
This one once carefully tuned by hand achieves 45ns on the ryzen but forces
a relax on i7 which costs a lot. It's only committed to be kept as an
illustration of what is possible.

Note that the l4 values are high, which is part of the optimization: some
closer cores manage to find their way more often than farther ones.

It looks like it should be possible to determine a self-uned loop count
inside the preamble without having to keep a variable like this. Also,
on the epilogue it is possible to use a ratio of avg_wait, with a bit
less performance.

Some variations of this model achieved extremely low l4 values (all <10)
for around 52-54ns. However it seems that it's finely tuned for the ryzen
only.