cas-bench: add two approaches to perform a load-exclusive
The idea here is to load the old value of a CAS using a load-exclusive, in
hope that CAS will rarely fail because once the value is read, few other
cores will be able to retrieve the line between the two.

Two approaches are attempted to perform the load-exclusive:
  - old = xadd(ptr, 0)
  - old = cmpxchg(ptr, dummy, dummy)

The first one fails 3 times less than with the regular load, and the second
one about 2 times less. The xadd is twice as slow on Ryzen, and 1.5 as slow
on core i7 (it must always write the result back), and the cmpxchg it only
1.5 times as slow on Ryzen and 10% slower on core i7 (only has to write
back if the dummy value matches).

There are still important differences between cores in number of
successful operations.