cas-bench: add more measurements in the code for o4r2
This adds some measures on ryzen and i7 when playing with write
threshold and unlock ratios. Ryzen and i7 are not optimized at the
same positions but it's still possible to get good results for both.
Both seem to favor reading more often but in fact it might be due to
not having the opportunity to write as often. This needs to be further
explored.

The loopcnt also counts the number of turns, but that's not really
equal since most of the time no slow operation is performed there. One
trick to be attempted would be to write to a shared global variable to
see the caches try to group their writes and possibly favor multiple
CPUs of the same core.