cas-bench: try a better exponential approach
Now we're updating and re-reading the avg_wait value every time we double
its number of loops, and are still reading the previous value at least
once every 32 loops.

This works pretty fine on i7, ryzen and armada but we should revisit
ratios because it will cause too much wait time on unlock.

On A55 it's still not good (100ns) while the unchecked version takes only 26.
We should improve this so that we start counting the loop only after it has
failed a few times.