cas-bench: experiment with going to the queue only after 3 failures
Here the idea is to avoid touching the queue entirely when not needed.
This significantly favors the A55 which is back to 30ns, without hurting
the other ones at all. It *does* create some unfairness in that a CPU
could theoretically starve the system, but the reality is that it will
also occasionally suffer from the other ones and will fail in turn.

Maybe keeping a local copy of the average loopcnt could significantly
help. In this case we could also do the same with failcnt and estimate
whether or not to take the test based on the latest one.