results: add thr-lat for EPYC 74F3
The machine has 24 cores set up as 8 groups of 3 cores each with its
own L3 cache. There seems to be two chiplets, each with 4 groups,
resulting in half of the cores seeing an even higher latency with
the other halves. Each L3 cache seems to be connected as a 3D torus.